+++ lib/elixir/test/elixir/kernel/warning_test.exs
@@ -11,36 +11,36 @@ defmodule Kernel.WarningTest do
   end
 
   test :unused_variable do
-    assert capture_err(fn ->
+    assert String.contains? capture_err(fn ->
       Code.eval_string """
       defmodule Sample do
         def hello(arg), do: nil
       end
       """
-    end) =~ "variable arg is unused"
+    end), "variable arg is unused"
   after
     purge Sample
   end
 
   test :unused_function do
-    assert capture_err(fn ->
+    assert String.contains? capture_err(fn ->
       Code.eval_string """
       defmodule Sample1 do
         defp hello, do: nil
       end
       """
-    end) =~ "function hello/0 is unused"
+    end), "function hello/0 is unused"
 
-    assert capture_err(fn ->
+    assert String.contains? capture_err(fn ->
       Code.eval_string """
       defmodule Sample2 do
         defp hello(0), do: hello(1)
         defp hello(1), do: :ok
       end
       """
-    end) =~ "function hello/1 is unused"
+    end), "function hello/1 is unused"
 
-    assert capture_err(fn ->
+    assert String.contains? capture_err(fn ->
       Code.eval_string %S"""
       defmodule Sample3 do
         def a, do: nil
@@ -49,32 +49,32 @@ defmodule Kernel.WarningTest do
         defp d(x), do: x
       end
       """
-    end) =~ "function c/2 is unused"
+    end), "function c/2 is unused"
   after
     purge [Sample1, Sample2, Sample3]
   end
 
   test :unused_cyclic_functions do
-    assert capture_err(fn ->
+    assert String.contains? capture_err(fn ->
       Code.eval_string """
       defmodule Sample do
         defp a, do: b
         defp b, do: a
       end
       """
-    end) =~ "function a/0 is unused"
+    end), "function a/0 is unused"
   after
     purge Sample
   end
 
   test :unused_macro do
-    assert capture_err(fn ->
+    assert String.contains? capture_err(fn ->
       Code.eval_string """
       defmodule Sample do
         defmacrop hello, do: nil
       end
       """
-    end) =~ "macro hello/0 is unused"
+    end), "macro hello/0 is unused"
   after
     purge Sample
   end
@@ -97,32 +97,32 @@ defmodule Kernel.WarningTest do
   end
 
   test :unused_default_args do
-    assert capture_err(fn ->
+    assert String.contains? capture_err(fn ->
       Code.eval_string %S"""
       defmodule Sample1 do
         def a, do: b(1, 2, 3)
         defp b(arg1 \\ 1, arg2 \\ 2, arg3 \\ 3), do: [arg1, arg2, arg3]
       end
       """
-    end) =~ "default arguments in b/3 are never used"
+    end), "default arguments in b/3 are never used"
 
-    assert capture_err(fn ->
+    assert String.contains? capture_err(fn ->
       Code.eval_string %S"""
       defmodule Sample2 do
         def a, do: b(1, 2)
         defp b(arg1 \\ 1, arg2 \\ 2, arg3 \\ 3), do: [arg1, arg2, arg3]
       end
       """
-    end) =~ "the first 2 default arguments in b/3 are never used"
+    end), "the first 2 default arguments in b/3 are never used"
 
-    assert capture_err(fn ->
+    assert String.contains? capture_err(fn ->
       Code.eval_string %S"""
       defmodule Sample3 do
         def a, do: b(1)
         defp b(arg1 \\ 1, arg2 \\ 2, arg3 \\ 3), do: [arg1, arg2, arg3]
       end
       """
-    end) =~ "the first default argument in b/3 is never used"
+    end), "the first default argument in b/3 is never used"
 
     assert capture_err(fn ->
       Code.eval_string %S"""
@@ -137,33 +137,33 @@ defmodule Kernel.WarningTest do
   end
 
   test :unused_import do
-    assert capture_err(fn ->
+    assert String.contains? capture_err(fn ->
       Code.compile_string """
       defmodule Sample do
         import :lists, only: [flatten: 1]
         def a, do: nil
       end
       """
-    end) =~ "unused import :lists"
+    end), "unused import :lists"
 
-    assert capture_err(fn ->
+    assert String.contains? capture_err(fn ->
       Code.compile_string """
       import :lists, only: [flatten: 1]
       """
-    end) =~ "unused import :lists"
+    end), "unused import :lists"
   after
     purge [Sample]
   end
 
   test :unused_alias do
-    assert capture_err(fn ->
+    assert String.contains? capture_err(fn ->
       Code.compile_string """
       defmodule Sample do
         alias :lists, as: List
         def a, do: nil
       end
       """
-    end) =~ "unused alias List"
+    end), "unused alias List"
   after
     purge [Sample]
   end
@@ -171,7 +171,7 @@ defmodule Kernel.WarningTest do
   test :unused_inside_dynamic_module do
     import List, only: [flatten: 1], warn: false
 
-    assert capture_err(fn ->
+    assert String.contains? capture_err(fn ->
       defmodule Sample do
         import String, only: [downcase: 1]
 
@@ -179,13 +179,13 @@ defmodule Kernel.WarningTest do
           flatten([1,2,3])
         end
       end
-    end) =~ "unused import String"
+    end), "unused import String"
   after
     purge [Sample]
   end
 
   test :unused_guard do
-    assert capture_err(fn ->
+    assert String.contains? capture_err(fn ->
       Code.eval_string """
       defmodule Sample1 do
         def is_atom_case do
@@ -197,9 +197,9 @@ defmodule Kernel.WarningTest do
         end
       end
       """
-    end) =~ "nofile:5: the guard for this clause evaluates to 'false'"
+    end), "nofile:5: the guard for this clause evaluates to 'false'"
 
-    assert capture_err(fn ->
+    assert String.contains? capture_err(fn ->
       Code.eval_string """
       defmodule Sample2 do
         def is_binary_cond do
@@ -211,20 +211,20 @@ defmodule Kernel.WarningTest do
         end
       end
       """
-    end) =~ "nofile:6: this clause cannot match because a previous clause at line 5 always matches"
+    end), "nofile:6: this clause cannot match because a previous clause at line 5 always matches"
   after
     purge [Sample1, Sample2]
   end
 
   test :unused_docs do
-    assert capture_err(fn ->
+    assert String.contains? capture_err(fn ->
       Code.eval_string """
       defmodule Sample1 do
         @doc "Oops"
         def hello
       end
       """
-    end) =~ "empty clause provided for nonexistent function or macro hello/0"
+    end), "empty clause provided for nonexistent function or macro hello/0"
   after
     purge [Sample1]
   end
@@ -251,33 +251,33 @@ defmodule Kernel.WarningTest do
   end
 
   test :clause_not_match do
-    assert capture_err(fn ->
+    assert String.contains? capture_err(fn ->
       Code.eval_string """
       defmodule Sample do
         def hello, do: nil
         def hello, do: nil
       end
       """
-    end) =~ "this clause cannot match because a previous clause at line 2 always matches"
+    end), "this clause cannot match because a previous clause at line 2 always matches"
   after
     purge Sample
   end
 
   test :clause_with_defaults_should_be_first do
-    assert capture_err(fn ->
+    assert String.contains? capture_err(fn ->
       Code.eval_string %S"""
       defmodule Sample do
         def hello(arg), do: nil
         def hello(arg \\ 0), do: nil
       end
       """
-    end) =~ "clause with defaults should be the first clause in def hello/1"
+    end), "clause with defaults should be the first clause in def hello/1"
   after
     purge Sample
   end
 
   test :unused_with_local_with_overridable do
-    assert capture_err(fn ->
+    assert String.contains? capture_err(fn ->
       Code.eval_string """
       defmodule Sample do
         def hello, do: world
@@ -286,7 +286,7 @@ defmodule Kernel.WarningTest do
         def hello, do: :ok
       end
       """
-    end) =~ "function world/0 is unused"
+    end), "function world/0 is unused"
   after
     purge Sample
   end
@@ -306,19 +306,19 @@ defmodule Kernel.WarningTest do
   end
 
   test :undefined_module_attribute do
-    assert capture_err(fn ->
+    assert String.contains? capture_err(fn ->
       Code.eval_string """
       defmodule Sample do
         @foo
       end
       """
-    end) =~ "undefined module attribute @foo, please remove access to @foo or explicitly set it to nil before access"
+    end), "undefined module attribute @foo, please remove access to @foo or explicitly set it to nil before access"
   after
     purge Sample
   end
 
   test :undefined_module_attribute_in_function do
-    assert capture_err(fn ->
+    assert String.contains? capture_err(fn ->
       Code.eval_string """
       defmodule Sample do
         def hello do
@@ -326,33 +326,33 @@ defmodule Kernel.WarningTest do
         end
       end
       """
-    end) =~ "undefined module attribute @foo, please remove access to @foo or explicitly set it to nil before access"
+    end), "undefined module attribute @foo, please remove access to @foo or explicitly set it to nil before access"
   after
     purge Sample
   end
 
   test :undefined_module_attribute_with_file do
-    assert capture_err(fn ->
+    assert String.contains? capture_err(fn ->
       Code.load_file(fixture_path("attribute_warning.ex"))
-    end) =~ "attribute_warning.ex:2: AttributeWarning (module) undefined module attribute @foo, please remove access to @foo or explicitly set it to nil before access"
+    end), "attribute_warning.ex:2: AttributeWarning (module) undefined module attribute @foo, please remove access to @foo or explicitly set it to nil before access"
   after
     purge AttributeWarning
   end
 
   test :in_guard_empty_list do
-    assert capture_err(fn ->
+    assert String.contains? capture_err(fn ->
       Code.eval_string """
       defmodule Sample do
         def a(x) when x in [], do: x
       end
       """
-    end) =~ "the guard for this clause evaluates to 'false'"
+    end), "the guard for this clause evaluates to 'false'"
   after
     purge Sample
   end
 
   test :no_effect_operator do
-    assert capture_err(fn ->
+    assert String.contains? capture_err(fn ->
       Code.eval_string """
       defmodule Sample do
         def a(x) do
@@ -361,13 +361,13 @@ defmodule Kernel.WarningTest do
         end
       end
       """
-    end) =~ "use of operator != has no effect"
+    end), "use of operator != has no effect"
   after
     purge Sample
   end
 
   test :undefined_function_for_behaviour do
-    assert capture_err(fn ->
+    assert String.contains? capture_err(fn ->
       Code.eval_string """
       defmodule Sample1 do
         use Behaviour
@@ -378,13 +378,13 @@ defmodule Kernel.WarningTest do
         @behaviour Sample1
       end
       """
-    end) =~ "undefined behaviour function foo/0 (for behaviour Sample1)"
+    end), "undefined behaviour function foo/0 (for behaviour Sample1)"
   after
     purge [Sample1, Sample2, Sample3]
   end
 
   test :undefined_macro_for_behaviour do
-    assert capture_err(fn ->
+    assert String.contains? capture_err(fn ->
       Code.eval_string """
       defmodule Sample1 do
         use Behaviour
@@ -395,13 +395,13 @@ defmodule Kernel.WarningTest do
         @behaviour Sample1
       end
       """
-    end) =~ "undefined behaviour macro foo/0 (for behaviour Sample1)"
+    end), "undefined behaviour macro foo/0 (for behaviour Sample1)"
   after
     purge [Sample1, Sample2, Sample3]
   end
 
   test :undefined_macro_for_protocol do
-    assert capture_err(fn ->
+    assert String.contains? capture_err(fn ->
       Code.eval_string """
       defprotocol Sample1 do
         def foo(subject)
@@ -410,13 +410,13 @@ defmodule Kernel.WarningTest do
       defimpl Sample1, for: Atom do
       end
       """
-    end) =~ "undefined protocol function foo/1 (for protocol Sample1)"
+    end), "undefined protocol function foo/1 (for protocol Sample1)"
   after
     purge [Sample1, Sample1.Atom]
   end
 
   test :overidden_def do
-    assert capture_err(fn ->
+    assert String.contains? capture_err(fn ->
       Code.eval_string """
       defmodule Sample do
         def foo(x, 1), do: x + 1
@@ -424,26 +424,26 @@ defmodule Kernel.WarningTest do
         def foo(x, 2), do: x * 2
       end
       """
-    end) =~ "nofile:4: clauses for the same def should be grouped together, def foo/2 was previously defined (nofile:2)"
+    end), "nofile:4: clauses for the same def should be grouped together, def foo/2 was previously defined (nofile:2)"
   after
     purge [Sample]
   end
 
   test :warning_with_overridden_file do
-    assert capture_err(fn ->
+    assert String.contains? capture_err(fn ->
       Code.eval_string """
       defmodule Sample do
         @file "sample"
         def foo(x), do: :ok
       end
       """
-    end) =~ "sample:3: variable x is unused"
+    end), "sample:3: variable x is unused"
   after
     purge [Sample]
   end
 
   test :typedoc_on_typep do
-    assert capture_err(fn ->
+    assert String.contains? capture_err(fn ->
       Code.eval_string """
       defmodule Sample do
         @typedoc "Something"
@@ -452,19 +452,19 @@ defmodule Kernel.WarningTest do
         def foo(), do: nil
       end
       """
-    end) =~ "nofile:3: type priv/0 is private, @typedoc's are always discarded for private types"
+    end), "nofile:3: type priv/0 is private, @typedoc's are always discarded for private types"
   after
     purge [Sample]
   end
 
   test :typedoc_with_no_type do
-    assert capture_err(fn ->
+    assert String.contains? capture_err(fn ->
       Code.eval_string """
       defmodule Sample do
         @typedoc "Something"
       end
       """
-    end) =~ "nofile:1: @typedoc provided but no type follows it"
+    end), "nofile:1: @typedoc provided but no type follows it"
   after
     purge [Sample]
   end