+++ lib/elixir/test/elixir/kernel_test.exs
@@ -11,23 +11,6 @@ defmodule KernelTest do
     refute bar!
   end
 
-  test :match do
-    assert ("abcd" =~ %r/c(d)/) == true
-    assert ("abcd" =~ %r/e/) == false
-
-    string = "^ab+cd*$"
-    assert (string =~ "ab+") == true
-    assert (string =~ "bb") == false
-
-    assert_raise FunctionClauseError, "no function clause matching in Kernel.=~/2", fn ->
-      1234 =~ "hello"
-    end
-
-    assert_raise FunctionClauseError, "no function clause matching in Kernel.=~/2", fn ->
-      1234 =~ %r"hello"
-    end
-  end
-
   test :^ do
     x = 1