+++ lib/elixir/test/elixir/regex_test.exs
@@ -8,12 +8,8 @@ defmodule Regex.BinaryTest do
     assert Regex.match?(%r/^b$/m, "a\nb\nc")
   end
 
-  test :precedence do
-    assert { "aa", :unknown } |> elem(0) =~ %r/(a)\1/
-  end
-
   test :backreference do
-    assert "aa" =~ %r/(a)\1/
+    assert Regex.match? %r/(a)\1/, "aa"
   end
 
   test :compile! do
@@ -34,7 +30,7 @@ defmodule Regex.BinaryTest do
 
   test :compile_with_erl_opts do
     { :ok, regex } = Regex.compile("foo\\sbar", [:dotall, {:newline, :anycrlf}])
-    assert "foo\nbar" =~ regex
+    assert Regex.match?(regex, "foo\nbar")
   end
 
   test :source do
@@ -46,8 +42,8 @@ defmodule Regex.BinaryTest do
   end
 
   test :unicode_by_default do
-    assert ("josé" =~ %r"\p{Latin}$")
-    refute ("£" =~ %r/\p{Lu}/) 
+    assert Regex.match?(%r"\p{Latin}$", "josé")
+    refute Regex.match?(%r/\p{Lu}/, "£")
   end
 
   test :groups do