+++ lib/ex_unit/lib/ex_unit/assertions.ex
@@ -160,10 +160,6 @@ defmodule ExUnit.Assertions do
     assert_operator :!=, left, right, "be not equal to (!=)"
   end
 
-  defp translate_assertion({ :=~, _, [left, right] }) do
-    assert_operator :=~, left, right, "match (=~)"
-  end
-
   defp translate_assertion({ :in, _, [left, right] }) do
     quote do
       left  = unquote(left)
@@ -187,14 +183,6 @@ defmodule ExUnit.Assertions do
           _ ->
             nil
       end
-    end
-  end
-
-  defp translate_assertion({ :!, _, [{ :=~, _, [left, right] }] }) do
-    quote do
-      left  = unquote(left)
-      right = unquote(right)
-      assert !(left =~ right), left, right, assertion: "match (=~)", negation: true
     end
   end
 
@@ -332,8 +320,8 @@ defmodule ExUnit.Assertions do
     error = assert_raise(exception, function)
 
     is_match = case message do
-      re  when is_regex(re)   -> error.message =~ re
-      bin when is_binary(bin) -> error.message == bin
+      re  when is_regex(re)   -> Regex.match?(re, error.message)
+      bin when is_binary(bin) -> bin == error.message
     end
 
     assert is_match, message, error.message,